requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
construct==2.5.3
cycler==0.10.0
Django==3.0.5
django-taggit==1.2.0
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
haystack==0.42
httplib2==0.17.0
idna==2.8
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
pandas==1.0.1
pefile==2019.4.18
Pillow==7.0.0
pipenv==2018.11.26
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-ptrace==0.9.5
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.7.9
virtualenv-clone==0.5.3
wrapt==1.11.2
